+++ src/main/java/growthcraft/grapes/common/compat/GrapesAspectRegistry.java
@@ -1,8 +1,6 @@
-package growthcraft.grapes;
+package growthcraft.grapes.common.compat;
 
-import growthcraft.grapes.common.items.ItemGrape;
 import growthcraft.grapes.shared.init.GrowthcraftGrapesItems;
-import net.minecraft.item.ItemStack;
 import thaumcraft.api.ThaumcraftApi;
 import thaumcraft.api.aspects.Aspect;
 import thaumcraft.api.aspects.AspectList;
